Add DTO

+++ src/main/java/board_project/board/service/BoardService.java
@@ -2,6 +2,9 @@ package board_project.board.service;
 
 import board_project.board.domain.Board;
 import board_project.board.domain.User;
+import board_project.board.dto.BoardResponseDto;
+import board_project.board.dto.BoardSaveRequestDto;
+import board_project.board.dto.BoardUpdateRequestDto;
 import board_project.board.repository.BoardRepository;
 import board_project.board.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -20,39 +23,56 @@ public class BoardService {
 
     // 게시글 작성
     @Transactional
-    public Long write(Long userId, String title, String content) {
-        User user = userRepository.findOne(userId);
-        Board board = Board.createBoard(title, content, user);
+    public Long write(BoardSaveRequestDto dto) {
+        User user = userRepository.findOne(dto.getUserId());
+        if(user == null) {
+            throw new IllegalStateException("해당 회원이 존재하지 않습니다. 회원 Id: " + dto.getUserId());
+        }
+        Board board = Board.createBoard(dto.getTitle(), dto.getContent(), user);
         boardRepository.save(board);
         return board.getId();
     }
 
     // 게시글 수정
     @Transactional
-    public void updateBoard(Long boardId, String title, String content) {
-        Board board = boardRepository.findOne(boardId);
-        board.updateBoard(title, content);
+    public void updateBoard(BoardUpdateRequestDto dto) {
+        Board board = boardRepository.findOne(dto.getBoardId());
+        if(board == null) {
+            throw new IllegalStateException("해당 게시글이 존재하지 않습니다. 게시글 Id : " + dto.getBoardId());
+        }
+        board.updateBoard(dto.getTitle(), dto.getContent());
     }
 
     @Transactional
     public void removeBoard(Long boardId) {
         Board board = boardRepository.findOne(boardId);
+        if(board == null) {
+            throw new IllegalStateException("해당 게시글이 존재하지 않습니다. 게시글 Id : " + boardId);
+        }
         board.remove();
         boardRepository.remove(board);
     }
 
     // 게시글 목록 조회
-    public List<Board> findBoards() {
-        return boardRepository.findAll();
+    public List<BoardResponseDto> findBoards() {
+        return boardRepository.findAll().stream()
+                .map(BoardResponseDto::new)
+                .toList();
     }
 
     // 게시글 단 건 조회
-    public Board findBoard(Long boardId) {
-        return boardRepository.findOne(boardId);
+    public BoardResponseDto findBoard(Long boardId) {
+        Board board = boardRepository.findOne(boardId);
+        if(board == null) {
+            throw new IllegalStateException("해당 게시글이 존재하지 않습니다. 게시글 Id : " + boardId);
+        }
+        return new BoardResponseDto(board);
     }
 
     // 작성자 별 게시글 조회
-    public List<Board> findBoardsByUser(Long userId) {
-        return boardRepository.findByUser(userId);
+    public List<BoardResponseDto> findBoardsByUser(Long userId) {
+        return boardRepository.findByUser(userId).stream()
+                .map(BoardResponseDto::new)
+                .toList();
     }
 }
